requirements.txt
flask
Flask-SQLAlchemy
pytest
requests
psycopg2-binary
